requirements.txt
flask
flask-cors
streamlit
requests
PyPDF2
python-docx
